+++ experiments/ClimaEarth/test/runtests.jl
@@ -7,6 +7,9 @@ gpu_broken = ClimaComms.device() isa ClimaComms.CUDADevice
 @safetestset "component model test: ClimaAtmos" begin
     include("component_model_tests/climaatmos_tests.jl")
 end
+@safetestset "component model test: ClimaLand integrated model" begin
+    include("component_model_tests/climaland_tests.jl")
+end
 @safetestset "component model test: prescr. sea ice" begin
     include("component_model_tests/prescr_seaice_tests.jl")
 end
@@ -24,4 +27,8 @@ end
 end
 @safetestset "AMIP test" begin
     include("amip_test.jl")
+end
+
+@safetestset "Restart test" begin
+    include("restart.jl")
 end